# docs/swagger.yaml
basePath: /api
definitions:
...
  create_channel.CreateChannelRequest:
    properties:
      isPrivate:
        type: boolean
      members:
        items:
          type: string
        type: array
      name:
        maxLength: 100
        minLength: 1
        type: string
      topic:
        type: string
    required:
    - name
    type: object
...
  github_com_supchat-lmrt_back-go_internal_workspace_member_usecase_invite_link_workspace_usecase_generate.CreateInviteLinkRequest:
    properties:
      workspaceId:
        type: string
    type: object
...
  list_channels.ListChannelResponse:
    properties:
      id:
        type: string
      name:
        type: string
      topic:
        type: string
    type: object
...
  models.UpdateJobRequest:
    properties:
      id:
        example: job-123
        type: string
      name:
        example: Développeur Senior
        type: string
    required:
    - id
    - name
    type: object
  models.UpdateJobResponse:
    properties:
      id:
        example: job-123
        type: string
      message:
        example: Rôle mis à jour avec succès
        type: string
      name:
        example: Développeur Senior
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  register.RegisterRequest:
    properties:
      password:
        minLength: 3
        type: string
      passwordConfirmation:
        type: string
      token:
        type: string
    required:
    - password
    - passwordConfirmation
    - token
    type: object
...
  validate.ValidateResetPasswordRequest:
    properties:
      newPassword:
        minLength: 8
        type: string
      newPasswordConfirmation:
        type: string
      token:
        type: string
    required:
    - newPassword
    - newPasswordConfirmation
    - token
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@supchat.io
    name: Support API SupChat
  description: API pour le service de messagerie SupChat
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SupChat API
  version: "1.0"
paths:
...
  /api/account/auth/register:
    post:
      consumes:
      - application/json
      description: Enregistre un nouvel utilisateur avec un token d'invitation et
        un mot de passe
      parameters:
      - description: Informations d'inscription
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/register.RegisterRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Inscription réussie
          schema:
            type: string
        "400":
          description: Paramètres invalides ou utilisateur déjà existant
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Inscription d'un utilisateur
      tags:
      - account
  /api/account/auth/token/access/renew:
    post:
      consumes:
      - application/json
      description: Génère un nouveau token d'accès à partir du token de rafraîchissement
        présent dans les cookies
      produces:
      - application/json
      responses:
        "204":
          description: Token d'accès renouvelé avec succès
          schema:
            type: string
        "401":
          description: Token de rafraîchissement manquant ou invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors du rafraîchissement du token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rafraîchissement du token d'accès
      tags:
      - account
...
  /api/account/reset-password/validate:
    post:
      consumes:
      - application/json
      description: Valide le token de réinitialisation et change le mot de passe de
        l'utilisateur
      parameters:
      - description: Données de réinitialisation du mot de passe
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validate.ValidateResetPasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Mot de passe réinitialisé avec succès
          schema:
            type: string
        "400":
          description: Paramètres invalides ou mots de passe non concordants
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la validation du token ou de la mise à jour
            du mot de passe
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validation de la réinitialisation de mot de passe
      tags:
      - account
...
  /api/job/{id}:
    delete:
      consumes:
      - application/json
      description: Supprime un rôle de travail existant en fonction de son ID
      parameters:
      - description: ID du rôle à supprimer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rôle supprimé avec succès
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Erreur de paramètre
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Supprimer un rôle de travail
      tags:
      - job
    put:
      consumes:
      - application/json
      description: Met à jour le nom d'un rôle de travail existant
      parameters:
      - description: ID du rôle à mettre à jour
        in: path
        name: id
        required: true
        type: string
      - description: Informations pour mettre à jour un rôle
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rôle mis à jour avec succès
          schema:
            $ref: '#/definitions/models.UpdateJobResponse'
        "400":
          description: Erreur de paramètre
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Non autorisé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mettre à jour un rôle de travail
      tags:
      - job
...
  /api/workspace-invite-link/create/{workspace_id}:
    post:
      consumes:
      - application/json
      description: Génère un lien d'invitation pour permettre à d'autres utilisateurs
        de rejoindre l'espace de travail
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Données de l'espace de travail
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_supchat-lmrt_back-go_internal_workspace_member_usecase_invite_link_workspace_usecase_generate.CreateInviteLinkRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Lien d'invitation généré
          schema:
            type: string
        "400":
          description: Requête invalide ou ID d'espace de travail manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour créer un lien d'invitation
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Espace de travail non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la génération du lien d'invitation
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Création d'un lien d'invitation
      tags:
      - workspace
      - invite
...
  /api/workspaces/{workspace_id}/channels:
    get:
      consumes:
      - application/json
      description: Retourne tous les canaux publics d'un espace de travail
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Liste des canaux
          schema:
            items:
              $ref: '#/definitions/list_channels.ListChannelResponse'
            type: array
        "400":
          description: ID de workspace manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la récupération des canaux
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Liste des canaux
      tags:
      - workspace
      - channel
    post:
      consumes:
      - application/json
      description: Crée un nouveau canal textuel dans un espace de travail spécifique
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Informations du canal à créer
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/create_channel.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Canal créé avec succès
          schema:
            type: string
        "400":
          description: Paramètres manquants ou invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour créer un canal
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la création du canal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Création d'un canal
      tags:
      - workspace
      - channel
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"